examples/bluetooth.html

<h2>Bluetooth Tizen-WRT</h2>

<body>

<input type="text" size=50 id="text">
<br>
<textarea cols=50 rows=20 id="output"></textarea>
<br>
<button id="button1">Click 1</button>
<button id="button2">Click 2</button>
<button id="button3">Click 3</button>
<button id="button4">Click 4</button>

<pre id="console"></pre>
<script src="js/js-test-pre.js"></script>
</body>

<script>

var output = document.getElementById("output");

var adapter = tizen.bluetooth.getDefaultAdapter();

var devices = [];
var device_index = 0;

function handle(button, text, callback) {
    var b = document.getElementById(button);
    b.innerText = text;
    b.addEventListener("click", callback);
}

var oldName = "";
var oldVisible = false;
function runChangePropertiesTest() {
    oldName = adapter.name;
    oldVisible = adapter.visible;

    debug("Adapter Name (oldName): " + adapter.name);
    debug("Adapter Visible (oldVisible): " + adapter.visible);

    shouldBe("adapter.name === oldName", "true");
    shouldBe("adapter.powered", "true");
    shouldBe("adapter.visible === oldVisible", "true");

    adapter.setName("FooBigBarrrrr", onChangedAdapterName, onErrorCallback);
    adapter.setVisible(!oldVisible, onChangedAdapterVisible, onErrorCallback);

    setTimeout(function() {
        shouldBe("adapter.name === oldName", "false");
    }, 1000);

    setTimeout(function() {
        adapter.setName(oldName, onChangedAdapterName, onErrorCallback);
        adapter.setVisible(!oldVisible, onChangedAdapterVisible, onErrorCallback);
    }, 1000);
};

var discoverDevicesSuccessCallback = {
    onstarted: function() {
        output.value += '\n\n## Discovery Started ##';
    },
    ondevicefound: function(device) {
        output.value += '\n' + device_index + '- Device Found: ' + device.name;
        devices[device_index] = device;
        device_index += 1;
    },
    ondevicedisappeared: function(address) {
        output.value += '\nDevice Disappeared: ' + address;
    },
    onfinished: function(devices) {
        output.value += '\n\n## Discovery Finished. Devices Found: ##';
        for (var i = 0; i < devices.length; i++) {
           output.value += '\nDevice Name: ' + devices[i].name;
        }
    }
};

var onErrorCallback = function(e) {
    console.log("\nFailed: " + e.message);
};

var onGotKnownDevices = function(devices) {
    output.value += '\n\n## Known Devices: ##';
    for (var i = 0; i < devices.length; i++) {
       output.value += '\nDevice Name: ' + devices[i].name;
    }
}

var onChangedAdapterName = function() {
    debug("Adapter New Name: " + adapter.name);
}

var onChangedAdapterVisible = function() {
    shouldBe("adapter.visible === !oldVisible", "true");
    oldVisible = adapter.visible;
}

handle("button1", "Scan", function() {
    adapter.discoverDevices(discoverDevicesSuccessCallback, function(e){
        console.log ("Failed to search devices: " + e.message + "(" + e.name + ")");
    });
});

handle("button2", "Stop Scan", function() {
    adapter.stopDiscovery(function() {
        output.value += "\n## Discovery Stopped ##";
        }, onErrorCallback);
});

handle("button3", "Get Known Devices", function() {
    adapter.getKnownDevices(onGotKnownDevices, onErrorCallback);
});

handle("button4", "Run Adapter properties test", function() {
    runChangePropertiesTest();
});
</script>
